### 2020-04-30 #36 @kneth
[Re] [Ten Year Challenge] Oscillating chemical reactions and phase separation simulated by molecular dynamics
**Original article:** https://aip.scitation.org/doi/abs/10.1063/1.475237
```
@article{Geisshirt:1997,
  author  = {Kenneth Geisshirt and Eigil L. Praestgaard and Soren Toxvaerd},
  title   = {Oscillating chemical reactions and phase separation simulated by molecular dynamics},
  journal = {Journal of Chemical Physics},
  pages   = {9406--9412},
  year    = {1997}
}
```

**Scientific domain:**  non-equilibrium thermodynamics, chemical kinetics
**Programming language:**  Fortran-77
