### 2020-04-30 #36 @kneth
Declaration of participation: https://github.com/ReScience/ten-years/issues/1#issuecomment-549156688
