+++ oldsrc/protocols/phd.sh
@@ -2,8 +2,8 @@
 
 rho=0.8
 
-mkdir density-${rho}
-cd density-${rho}
+mkdir phd-${rho}
+cd phd-${rho}
 
 for T in 0.4 0.6 0.7 0.8 0.9 1.0 1.1 1.2 1.3 1.4 1.5 1.6 1.7 1.8 2.0 2.1 2.2 2.3 2.4 2.5 2.6 2.7 2.8 2.9 3.0 3.2 3.4 3.6 3.8 4.0 4.33 4.66 5.0 5.33 5.66 6.0 6.5 7.0 7.5 8.0 9.0 10.0
 do
@@ -16,7 +16,7 @@ nY              256
 nZ              0
 rho             ${rho}
 rcut1           2.5
-rcut2           2.5
+rcut2           1.122462
 temperature     ${T}
 nsteps          250000
 probability 1   0.001
