### 2020-06-04 #36 @fhoefling
I started with the review. Some first feedback on the software:

A detailed guide of how to proceed is missing and should be given. Yet, building the programme went smoothly, just by installing `gfortran` and running `make` in `oldsrc`.

Next, I started the script `phd.sh` from the `protocols` folder. On a Xeon E3-1241 @ 3.50GHz, the script was running for more than 5 hours on a single core and created almost 8 GB of data. Is this the intended behaviour? If so they user should be alerted. No further output was created.

I've got the following Python error repeatedly:

> File "../../../diffcoeff.py", line 8, in \<module\>
      from sklearn.linear_model import LinearRegression
      ModuleNotFoundError: No module named 'sklearn'

Please use the linear regression functionality from a more basic library. Machine learning approaches were not needed in the original work.

The provided README is rather brief. It mentions some means to run the programme in parallel. Is this still possible today?
