### 2020-06-12 #36 @fhoefling
Actually, I can only continue with the reviewing after the aforementioned
issues have been solved. Currently, the programme does not output data. I
apologise that I have not communicated this more clearly.
