### 2020-08-11 #36 @pdebuyl
@kneth could you address the comments?
