### 2020-08-23 #36 @kneth
@pdebuyl

Sorry for not getting things done earlier. I have created https://github.com/kneth/MDreac/pull/4 to address the first comments. I hope to rerun the analysis steps once more to see if I get the same result as before.
